+++ packages/liveblocks-schema/src/checker/index.ts
@@ -1,19 +1,19 @@
 import type {
-  ArrayExpr,
+  ArrayType,
   Definition,
   Document,
   FieldDef,
   Identifier,
-  LiveMapExpr,
-  LiveStructureExpr,
-  ObjectLiteralExpr,
+  LiveMapType,
+  LiveType,
+  ObjectLiteralType,
   ObjectTypeDefinition,
   Range,
-  TypeExpr,
+  Type,
   TypeName,
   TypeRef,
 } from "../ast";
-import { isBuiltInScalar, isLiveStructureExpr, visit } from "../ast";
+import { isLiveType, isScalarType, visit } from "../ast";
 import { assertNever } from "../lib/assert";
 import { didyoumean as dym } from "../lib/didyoumean";
 import type { ErrorReporter, Suggestion } from "../lib/error-reporting";
@@ -26,39 +26,21 @@ const TYPENAME_REGEX = /^[A-Z_]/;
 
 // TODO Ideally _derive_ this list of builtins directly from the grammar
 // instead somehow?
-const BUILTINS = ["String", "Int", "Float", "Boolean"] as const;
+const BUILTINS = ["string", "number", "boolean", "null"] as const;
 
 function suggest_general(value: string, alternatives: string[]): string[] {
   // Never suggest "Storage"
   alternatives = alternatives.filter((key) => key !== "Storage");
-
-  const suggestions = dym(value, alternatives);
-
-  // Special hack:
-  // It can be expected that people will try to put "number" in as a type,
-  // because that's TypeScript's syntax. If there is no custom type name found
-  // that closely matches this typo, then try to suggest one more thing to
-  // nudge them. (But only if Float and Int are already legal suggestions.)
-  if (
-    suggestions.length === 0 &&
-    alternatives.includes("Float") &&
-    alternatives.includes("Int")
-  ) {
-    if (/^num(ber)?$/i.test(value)) {
-      return ["Float", "Int"];
-    }
-  }
-
-  return suggestions;
+  return dym(value, alternatives);
 }
 
 /**
  * Whether a type expression is a generic LiveXxx<> type.
  */
-function isLiveStructure(node: TypeExpr): node is LiveStructureExpr | TypeRef {
+function isLiveStructure(node: Type): node is LiveType | TypeRef {
   return (
     // LiveList<...>, or LiveMap<...>
-    isLiveStructureExpr(node) ||
+    isLiveType(node) ||
     // e.g. LiveObject<...>
     (node._kind === "TypeRef" && node.asLiveObject)
   );
@@ -77,7 +59,8 @@ function didyoumeanify(message: string, alternatives: string[]): string {
 /**
  * Reserve these names for future use.
  */
-const RESERVED_TYPENAMES_REGEX = /^Live|^(Presence|Array)$/i;
+const RESERVED_TYPENAMES_REGEX =
+  /^live|^(presence|array|string|int|float|number|boolean|null)$/i;
 
 /**
  * Reserve these identifiers for future use.
@@ -195,18 +178,14 @@ function checkNoDuplicateFields(fieldDefs: FieldDef[], context: Context): void {
   }
 }
 
-function ensureNoLiveStructure(
-  expr: TypeExpr,
-  where: string,
-  context: Context
-) {
+function ensureNoLiveStructure(expr: Type, where: string, context: Context) {
   if (isLiveStructure(expr)) {
     context.report(`Cannot use Live construct ${where}`, expr.range);
   }
 }
 
-function checkObjectLiteralExpr(
-  obj: ObjectLiteralExpr,
+function checkObjectLiteralType(
+  obj: ObjectLiteralType,
   context: Context
 ): void {
   checkNoDuplicateFields(obj.fields, context);
@@ -217,11 +196,18 @@ function checkObjectLiteralExpr(
   }
 }
 
-function checkArrayExpr(arr: ArrayExpr, context: Context): void {
+function checkArrayType(arr: ArrayType, context: Context): void {
   ensureNoLiveStructure(arr.ofType, "inside an array", context);
 }
 
 function checkTypeName(node: TypeName, context: Context): void {
+  if (BUILTINS.some((bname) => bname === node.name)) {
+    context.report(
+      `Name ${quote(node.name)} is a built-in and cannot be redefined`,
+      node.range
+    );
+  }
+
   if (!TYPENAME_REGEX.test(node.name)) {
     context.report(
       "Type names should start with an uppercase character",
@@ -229,18 +215,13 @@ function checkTypeName(node: TypeName, context: Context): void {
     );
   }
 
-  // Continue collecting more errors
-
-  if (BUILTINS.some((bname) => bname === node.name)) {
-    context.report(
-      `Name ${quote(node.name)} is a built-in and cannot be redefined`,
-      node.range
-    );
-  } else if (RESERVED_TYPENAMES_REGEX.test(node.name)) {
-    context.report(
-      `Name ${quote(node.name)} is reserved for future use`,
-      node.range
-    );
+  if (!context.errorReporter.hasErrors) {
+    if (RESERVED_TYPENAMES_REGEX.test(node.name)) {
+      context.report(
+        `Name ${quote(node.name)} is reserved for future use`,
+        node.range
+      );
+    }
   }
 }
 
@@ -324,11 +305,11 @@ function checkTypeRef(ref: TypeRef, context: Context): void {
 }
 
 function checkNoForbiddenRefs(
-  node: ObjectTypeDefinition | TypeExpr,
+  node: ObjectTypeDefinition | Type,
   context: Context,
   forbidden: Set<string>
 ): void {
-  if (isBuiltInScalar(node)) {
+  if (isScalarType(node)) {
     return;
   }
 
@@ -344,7 +325,7 @@ function checkNoForbiddenRefs(
       }
       break;
 
-    case "ObjectLiteralExpr":
+    case "ObjectLiteralType":
       for (const field of node.fields) {
         // TODO for later. Allow _some_ self-references. For example, if
         // `field.optional`, then it'd be perfectly fine to use
@@ -355,12 +336,12 @@ function checkNoForbiddenRefs(
       }
       break;
 
-    case "ArrayExpr":
-    case "LiveListExpr":
+    case "ArrayType":
+    case "LiveListType":
       checkNoForbiddenRefs(node.ofType, context, forbidden);
       break;
 
-    case "LiveMapExpr":
+    case "LiveMapType":
       checkNoForbiddenRefs(node.keyType, context, forbidden);
       checkNoForbiddenRefs(node.valueType, context, forbidden);
       break;
@@ -425,13 +406,13 @@ function checkObjectTypeDefinition(
   checkNoForbiddenRefs(def, context, new Set([def.name.name]));
 }
 
-function checkLiveMapExpr(node: LiveMapExpr, context: Context): void {
+function checkLiveMapType(node: LiveMapType, context: Context): void {
   // Check that the `keyType` is always `String`, never another type
   if (node.keyType._kind !== "StringType") {
     context.report(
-      "Only 'String' keys are currently supported in LiveMaps",
+      "Only 'string' keys are currently supported in LiveMaps",
       node.keyType.range,
-      [{ type: "replace", name: "String" }]
+      [{ type: "replace", name: "string" }]
     );
   }
 }
@@ -487,12 +468,12 @@ function decideStaticOrLive(doc: Document, context: Context): void {
       visit(
         def,
         {
-          LiveListExpr: () => {
+          LiveListType: () => {
             liveObjRefs.set(def.name.name, null);
             throw "break";
           },
 
-          LiveMapExpr: () => {
+          LiveMapType: () => {
             liveObjRefs.set(def.name.name, null);
             throw "break";
           },
@@ -611,10 +592,10 @@ export function check(
   visit(
     doc,
     {
-      ArrayExpr: checkArrayExpr,
+      ArrayType: checkArrayType,
       Identifier: checkIdentifier,
-      LiveMapExpr: checkLiveMapExpr,
-      ObjectLiteralExpr: checkObjectLiteralExpr,
+      LiveMapType: checkLiveMapType,
+      ObjectLiteralType: checkObjectLiteralType,
       ObjectTypeDefinition: checkObjectTypeDefinition,
       TypeName: checkTypeName,
       TypeRef: checkTypeRef,