+++ CHANGELOG.md
@@ -1,5 +1,22 @@
-# 0.0.11-beta2
+# 0.0.11-beta3
 
+- Make all built-in types lowercased, following TypeScript syntax more closely
+  instead of GraphQLs:
+  - `String` → `string`
+  - `Float` → `number`
+  - `Boolean` → `boolean`
+  - Removed `Int` type. It will get reintroduced later in a better way.
+  - Added a new `null` type. (It's not super useful yet, because there is no
+    union construct yet, but that will be added soon.)
+- Added parser option to parser to still allow legacy built-ins (to be able to
+  parse existing valid schemas created before this breaking change)
+- The following AST nodes have been renamed:
+  - `ArrayExpr` -> `ArrayType`
+  - `BuiltInScalar` -> `ScalarType`
+  - `LiveListExpr` -> `LiveListType`
+  - `LiveMapExpr` -> `LiveMapType`
+  - `ObjectLiteralExpr` -> `ObjectLiteralType`
+  - `TypeExpr` -> `Type`
 - Add CodeMirror integration
 
 # 0.0.10